Create EnumFlags Element Refactor Boolean Node

+++ Assets/Scripts/Plugins/AmuletGames/DialogueSystem/Engine/Models/Components/Modifiers/EventModifier/EventModifierModel.cs
@@ -21,7 +21,7 @@ namespace AG.DS
         /// <summary>
         /// The modifier's delay seconds value.
         /// </summary>
-        [SerializeField] public double DelaySeconds;
+        [SerializeField] public float DelaySeconds;
 
 
         /// <summary>
